requirements.txt
streamlit==1.41.1
pypdf==5.0.1
python-dotenv==1.0.1
requests==2.32.3
jsonschema==4.23.0
pytest==8.3.4
fastapi==0.115.6
uvicorn==0.30.6
jinja2==3.1.5
